requirements.txt
discord.py
aiohttp
google-api-python-client
bs4
psutil
psycopg2